fix: two mypy failures
These show up when you run `mypy` by itself, but not during pre-commit.
They also showed up in this failed CI job:

https://github.com/ApeWorX/eip712/runs/2301435619

    eip712/messages.py:5: error: Skipping analyzing 'eth_account._utils.structured_data.hashing': found module but no type hints or library stubs
    eip712/messages.py:5: note: See https://mypy.readthedocs.io/en/latest/running_mypy.html#missing-imports
    eip712/messages.py:50: error: Argument 1 to "fields" has incompatible type "EIP712Type"; expected "Type[Any]"
    eip712/messages.py:80: error: Argument 1 to "fields" has incompatible type "EIP712Message"; expected "Type[Any]"

For "found module but no type hints", I don't see types anywhere for
this, so we'll just ignore it for now.

For the two "incompatible type" errors, `fields()` takes a class as its
first argument, not an instance.